requirements.txt
google-generativeai
streamlit
python-dotenv
google
pillow
